feat: 获取商品列表接口 50%

+++ main/app/dto/common_resp.go
@@ -8,9 +8,9 @@ type Response struct {
 
 // 分页响应
 type PageResponse struct {
-	PageNo   int `json:"page_no"`   // 当前页码
-	PageSize int `json:"page_size"` // 每页大小
-	Total    int `json:"total"`     // 总数
+	PageNo   int   `json:"page_no"`   // 当前页码
+	PageSize int   `json:"page_size"` // 每页大小
+	Total    int64 `json:"total"`     // 总数
 }
 
 // 语言响应
